Descentraliza factories
