+++ sistemita/core/tests/factories/__init__.py
@@ -1,0 +1,3 @@
+from .clientes import *  # noqa
+from .mediopagos import *  # noqa
+from .proveedores import *  # noqa
